[data_injection] - fixed/improved install/upgrade process

+++ front/plugin_data_injection.config.form.php
@@ -39,33 +39,31 @@ if (!defined('GLPI_ROOT')) {
 $NEEDED_ITEMS=array("data_injection");
 include (GLPI_ROOT."/inc/includes.php");
 
-if(!isset($_SESSION["glpi_plugin_data_injection_installed"]) || $_SESSION["glpi_plugin_data_injection_installed"]!=1) {
-	
-	commonHeader($LANG["common"][12],$_SERVER['PHP_SELF'],"config","plugins");
-	
-			echo "<div align='center'>";
-			echo "<table class='tab_cadre' cellpadding='5'>";
-			echo "<tr><th>".$DATAINJECTIONLANG["setup"][1]."</th></tr>";
-			echo "<tr class='tab_bg_1'><td align='center'><a href='http://glpi-project.org/wiki/doku.php?id=".substr($_SESSION["glpilanguage"],0,2).":plugins:data_injection_use' target='_blank'>".$DATAINJECTIONLANG["setup"][11]."</a></td></tr>";
-			echo "<tr class='tab_bg_1'><td align='center'>";
-			echo "<a href='plugin_data_injection.install.php'>".$DATAINJECTIONLANG["setup"][3]."</a></td></tr>";
-			echo "</table></div>";
+commonHeader($LANG["common"][12],$_SERVER['PHP_SELF'],"config","plugins");
+
+echo "<div align='center'>";
+echo "<table class='tab_cadre' cellpadding='5'>";
+echo "<tr><th>".$DATAINJECTIONLANG["setup"][1]."</th></tr>";
+echo "<tr class='tab_bg_1'><td align='center'><a href='http://glpi-project.org/wiki/doku.php?id=".substr($_SESSION["glpilanguage"],0,2).":plugins:data_injection_use' target='_blank'>".$DATAINJECTIONLANG["setup"][11]."</a></td></tr>";
+
+if(isset($_SESSION["glpi_plugin_data_injection_installed"]) && $_SESSION["glpi_plugin_data_injection_installed"]>0) {
+	// If installed => configure
+	if (haveRight("profile","w")){
+		echo "<tr class='tab_bg_1'><td align='center'><a href=\"../front/plugin_data_injection.profile.php\">".$DATAINJECTIONLANG["setup"][9]."</a></td/></tr>";
+	}
+} else {
+	// If not installed => install or upgrade
+	echo "<tr class='tab_bg_1'><td align='center'>";
+	echo "<a href='plugin_data_injection.install.php'>".$DATAINJECTIONLANG["setup"][3]."</a></td></tr>";		
 }
-else
-{
-	commonHeader($DATAINJECTIONLANG["config"][1], $_SERVER["PHP_SELF"],"plugins","data_injection");
-		echo "<div align='center'>";
-		echo "<table class='tab_cadre' cellpadding='6'>";
-		echo "<tr class='tab_bg_2'><th>" . $DATAINJECTIONLANG["setup"][1]."</th></tr>";
-		if (haveRight("profile","w")){
-			echo "<tr class='tab_bg_1'><td align='center'><a href=\"../front/plugin_data_injection.profile.php\">".$DATAINJECTIONLANG["setup"][9]."</a></td/></tr>";
-		}
-		echo "<tr class='tab_bg_1'><td align='center'><a href='http://glpi-project.org/wiki/doku.php?id=".substr($_SESSION["glpilanguage"],0,2).":plugins:data_injection_use' target='_blank'>".$DATAINJECTIONLANG["setup"][11]."</a></td></tr>";
-		if (haveRight("config","w")){
-			echo "<tr class='tab_bg_1'><td align='center'><a href=\"../front/plugin_data_injection.uninstall.php\">".$DATAINJECTIONLANG["setup"][5]."</a></td/></tr>";
-		}
-		echo "</table></div>";
+if (TableExists("glpi_plugin_data_injection_models") && haveRight("config","w")){
+	// If installed (or need upgrade) => uninstall
+
+	echo "<tr class='tab_bg_1'><td align='center'><a href=\"../front/plugin_data_injection.uninstall.php\">".$DATAINJECTIONLANG["setup"][5]."</a></td/></tr>";
 }
+	
+
+echo "</table></div>";
 
 commonFooter();
 ?>
